PittsburghEventsFetcher/requirements.txt
arrow==1.3.0
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
ics==0.7.2
idna==3.7
numpy==2.0.0
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
TatSu==5.8.3
tk==0.1.0
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
urllib3==2.2.2
